PC-1227: restructure routing files
leaves space for the backwards routing to be its own file
