+++ tests/routing/test_forwards_routing.py
@@ -1,165 +1,8 @@
 import pytest
 
-from help_to_heat.frontdoor.routing import (
-    address_choice_field,
-    address_choice_field_enter_manually,
-    address_choice_field_write_address,
-    address_manual_page,
-    address_page,
-    address_select_choice_field,
-    address_select_choice_field_enter_manually,
-    address_select_choice_field_select_address,
-    address_select_page,
-    benefits_field,
-    benefits_page,
-    bulb_warning_page,
-    confirm_and_submit_page,
-    contact_details_page,
-    council_tax_band_field,
-    council_tax_band_field_a,
-    council_tax_band_field_b,
-    council_tax_band_field_c,
-    council_tax_band_field_d,
-    council_tax_band_field_e,
-    council_tax_band_field_f,
-    council_tax_band_field_g,
-    council_tax_band_field_h,
-    council_tax_band_field_i,
-    council_tax_band_page,
-    council_tax_field_bands,
-    country_field,
-    country_field_england,
-    country_field_northern_ireland,
-    country_field_scotland,
-    country_field_wales,
-    country_page,
-    duplicate_uprn_field,
-    epc_accept_suggested_epc_field,
-    epc_found_field,
-    epc_ineligible_page,
-    epc_page,
-    epc_rating_is_eligible_field,
-    epc_select_choice_field,
-    epc_select_choice_field_enter_manually,
-    epc_select_choice_field_epc_api_fail,
-    epc_select_choice_field_select_epc,
-    epc_select_page,
-    field_dont_know,
-    field_no,
-    field_yes,
-    get_next_page,
-    household_income_field,
-    household_income_field_less_than_threshold,
-    household_income_field_more_than_threshold,
-    household_income_page,
-    loft_access_field,
-    loft_access_field_no,
-    loft_access_field_yes,
-    loft_access_page,
-    loft_field,
-    loft_field_no,
-    loft_field_yes,
-    loft_insulation_field,
-    loft_insulation_field_dont_know,
-    loft_insulation_field_less_than_threshold,
-    loft_insulation_field_more_than_threshold,
-    loft_insulation_page,
-    loft_page,
-    northern_ireland_ineligible_page,
-    number_of_bedrooms_field,
-    number_of_bedrooms_field_one,
-    number_of_bedrooms_field_studio,
-    number_of_bedrooms_field_three_or_more,
-    number_of_bedrooms_field_two,
-    number_of_bedrooms_page,
-    own_property_field,
-    own_property_field_landlord,
-    own_property_field_own_property,
-    own_property_field_social_housing,
-    own_property_field_tenant,
-    own_property_fields_non_social_housing,
-    own_property_page,
-    park_home_field,
-    park_home_ineligible_page,
-    park_home_main_residence_field,
-    park_home_main_residence_page,
-    park_home_page,
-    property_ineligible_page,
-    property_subtype_field,
-    property_subtype_field_detached,
-    property_subtype_field_end_terrace,
-    property_subtype_field_ground_floor,
-    property_subtype_field_middle_floor,
-    property_subtype_field_semi_detached,
-    property_subtype_field_terraced,
-    property_subtype_field_top_floor,
-    property_subtype_page,
-    property_type_field,
-    property_type_field_apartment,
-    property_type_field_bungalow,
-    property_type_field_house,
-    property_type_field_park_home,
-    property_type_page,
-    referral_already_submitted_page,
-    schemes_page,
-    success_page,
-    summary_page,
-    supplier_field,
-    supplier_field_british_gas,
-    supplier_field_bulb,
-    supplier_field_e,
-    supplier_field_edf,
-    supplier_field_eon_next,
-    supplier_field_foxglove,
-    supplier_field_octopus,
-    supplier_field_ovo,
-    supplier_field_scottish_power,
-    supplier_field_shell,
-    supplier_field_utilita,
-    supplier_field_utility_warehouse,
-    supplier_page,
-    utility_warehouse_warning_page,
-    wall_insulation_field,
-    wall_insulation_field_dont_know,
-    wall_insulation_field_no,
-    wall_insulation_field_some,
-    wall_insulation_field_yes,
-    wall_insulation_page,
-    wall_type_field,
-    wall_type_field_cavity,
-    wall_type_field_dont_know,
-    wall_type_field_mix,
-    wall_type_field_not_listed,
-    wall_type_field_solid,
-    wall_type_page,
-)
-
-flow_park_home = "park home flow"
-flow_main = "main (not park) home flow"
-flow_social_housing = "social housing flow"
-
-
-def _get_park_home_flow_answers():
-    for own_property in own_property_fields_non_social_housing:
-        yield {own_property_field: own_property, park_home_field: field_yes}
-
-
-def _get_main_flow_answers():
-    for own_property in own_property_fields_non_social_housing:
-        yield {own_property_field: own_property, park_home_field: field_no}
-
-
-def _get_social_housing_flow_answers():
-    yield {own_property_field: own_property_field_social_housing, park_home_field: field_no}
-
-
-def _get_flow_answers(flow):
-    if flow == flow_park_home:
-        return list(_get_park_home_flow_answers())
-    if flow == flow_main:
-        return list(_get_main_flow_answers())
-    if flow == flow_social_housing:
-        return list(_get_social_housing_flow_answers())
+from help_to_heat.frontdoor.consts import *
+from help_to_heat.frontdoor.routing.forwards_routing import get_next_page
+from tests.routing import flow_park_home, flow_main, flow_social_housing, _get_flow_answers
 
 
 @pytest.mark.parametrize(
